requirements.txt
diffusers~=0.33.1
numpy~=1.26.4
pillow~=11.1.0
fastapi~=0.115.12
pydantic~=2.11.3
psutil~=7.0.0
pydantic-settings~=2.8.1
uvicorn~=0.34.0
transformers~=4.51.3
accelerate~=1.6.0
peft~=0.15.2
--index-url https://download.pytorch.org/whl/cu124
torch==2.6.0+cu124
torchvision==0.21.0+cu124
torchaudio==2.6.0
controlnet_aux==0.0.9
clip-interrogator==0.6.0